feat(content): channels hub at /dashboard/content
Today /dashboard/content lands directly on the Beehiiv-specific library.
With Substack, Kit, LinkedIn, Meta, Google Ads, and friends arriving,
that page needs to become a hub of content channels grouped by category,
with the Beehiiv library moved to /dashboard/content/beehiiv (unchanged).

Changes:
- Move existing page.tsx, columns.tsx, [id]/page.tsx into ./beehiiv/
  (zero edits to grid, filters, formatting — pure rename via git mv)
- New page.tsx: tabbed channels hub (All + Newsletters / Social /
  Web & SEO / Ads / Messaging) using the @shadcnblocks/settings-integrations4
  pattern, fed from a single channels.config.ts registry.
- channels.config.ts: 18 channels seeded (live: Beehiiv only; available:
  LinkedIn / X / Facebook / Instagram / Meta Ads / WhatsApp; coming_soon:
  Substack / Kit / etc.). Adding a channel = one row here.
- Hub auto-redirects to the user's only live channel when exactly one is
  connected — single-channel users skip the extra click.
- Per-row health line ("Last synced 2h ago" / red "Last synced 30 days
  ago — check connection") reads `int_connections.lastSyncAt` so the
  Beehiiv sync incident pattern is visible at a glance.
- Update internal links: /dashboard/content/<id> → /dashboard/content/beehiiv/<id>
  in the table row click handlers and content-ref.tsx hover-card link.
- Update detail-page back button: /dashboard/content → /dashboard/content/beehiiv.

Sidebar / command-menu / overview tile links to /dashboard/content still
work — they hit the new hub, which routes single-channel users straight
through to Beehiiv.

Block install also added settings-integrations4.tsx + scrollable-tabslist.tsx
under src/components/. The hub uses scrollable-tabslist directly; the
example block stays as a reference for future hub variants.
